Add useFetchLeaderboardData.tsx
