+++ packages/client/src/pages/Leaderboard/useFetchLeaderboardData/index.ts
@@ -1,0 +1,1 @@
+export * from './useFetchLeaderboardData';
